scan and apply_layers: milestone 1
This commit adds the lowering of scan to HLO While op. It also
introduce apply_layers which can sequentially apply a bunch of layers
using scan underneath.

In this milestone we use AOTAutograd to obtain the backward of the
function being scanned. Users can either save the activations in
fn or recompute them by passing different graph partitioners to
AOTAutograd.